Add logic to filter commits by a user-defined filter_proc. This is the first step in filtering by the "unapproved" status.
If the first round of commits didn't produce enough positive results according to filter_proc,
more pages will be fetched until we have as many commits as we're looking for.

The next step is to ensure that all searches are time-bound since this search method needs to have some well-defined
end point where it can give up paging through commits.