Fix a test bug by not requiring test_helper directly.
You need to require (unit|integration)_test_helper so that pathological
is loaded and the paths are set up.

This only started breaking recently because of a naming change which
changed the order that the unit tests were run. If a unit test with
the proper require is loaded first, then pathological will already be
in the environment for subsequent test runs.